serialization -> refactor
